+++ packages/lit-html/CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Added `renderBefore` to render options. If specified, content is rendered before the node given via render options, e.g. `{renderBefore: node}`.
 * Added development mode, which can be enabled by setting the `development` Node exports condition. See `README.md` for more details.
+- Added `unsafeStatic()`, which allows template authors to add strings to the
+  static structure of the template, before it's parsed as HTML.
 
 ### Fixed
 * All usage of `instanceof` has been removed, making rendering more likely to